requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.28.1
python-dotenv==1.0.1
openai==1.53.0
pydantic==2.9.0
tenacity==8.2.3
aiobreaker==1.2.0
prometheus_client==0.23.1